Fix use of uninterpreted constants in smp.sal
The use of 'uninterp_msg' function was producing only a single
concrete constant of value 'faulty', but the desired effect was to
get uninterpreted message values in the faulty case.

The theorems & lemmas from previous commit go through with only a
few simple modifications. The one notable addition is the lemma
'afault' which seems to also be neccessary in larger examples.