Add a 'basic types' section with option to select a finite model
The top most section of the system module now has two sections to
choose from (by commenting in or out). One choice is the more general
case where time is unbounded and REAL and messages are INTEGERs, and
faults are uninterpreted integer constants. The other choice is case
where time, messages, and the faults are all integers in a finite range.

This is useful for running diagnostics over the model, e.g.
'sal-deadlock-checker' and 'sal-path-finder'.